Release v0.2.0
Modify behavior during UpdateDR. Now the script no longer checks for differences or absence of LIF or route to the destination during an UpdateDR (or Update-SvmDr).
This is in order to preserve any differences in gateway or other between the Primary and the Secondary.
If you need to modify/create your LIFs or Routes configuration on the Secondary, you must now do so only with a ConfigureDR (or New-SvmDr)
Change version display and align with Github Release number